adding TcpInstance to keep track of each connection
Connecting, disconnecting and trying to connect to a different
connection causes both connections to close when the first one fails to
connect.

Adding a class to keep track of the current TpcClient and threads.
If the connection is closed while Connect is still blocking the thread
then SocketException will be throw before ThreadInterruptedException
is caught