fix: Do not set timeouts on listener
On linux setting the recv timeout will cause the blocking Accept call to timeout with EACCEPT (which mono interprets as EWOULDBLOCK). See https://stackoverflow.com/questions/1917814/eagain-error-for-accept-on-blocking-socket/1918118#1918118
I'm not sure if the send/nodelay would cause any issues, but we set the timeouts on accepted sockets anyways, so this is unnecessary.